script/new/good/requirements.txt
pyautogui
psutil
pystray
Pillow
keyboard
wmi
configparser
